final version but changes needs to be made. 1. formatting into modules 2. serialization and deserialization, 3. new model for books where i need to create a new table in db for book_counts 4. new db to store token validation

+++ app/routes.py
@@ -1,23 +1,32 @@
 from utils import generate_hashed_password
 from constants import book_properties, book_allowed_fields, member_properties, member_allowed_properties, \
     ERROR_CONSTANTS
-from helper import is_record_available, response_formatter, get_missing_keys, get_error_details, map_keys_to_values
+from helper import is_record_available, response_formatter, get_missing_keys, get_error_details, map_keys_to_values, \
+    get_email_from_jwt, is_admin
 from model import Books, Db, Member, BookMember
 from flask import Flask, request, jsonify
 from _datetime import datetime
 from sqlalchemy import func
 
 
+def get_params_details():
+    page = request.args.get("page", default=1, type=int)
+    per_page = request.args.get("per_page", default=3, type=int)
+    return page, per_page
+
+
 def get_books():
     try:
-        books = Books.query.all()
+        page, per_page = get_params_details()
         author, title = request.args.get("author"), request.args.get("title")
+        books = Books.query.paginate(page=page, per_page=per_page)
         if author and title:
-            books = Books.query.filter(Books.author.like(author + "%")).filter(Books.title.like(title + "%")).all()
+            books = Books.query.filter(Books.author.like(author + "%")).filter(Books.title.like(title + "%")).paginate(
+                page=page, per_page=per_page)
         elif author:
-            books = Books.query.filter(Books.author.like(author + "%")).all()
+            books = Books.query.filter(Books.author.like(author + "%")).paginate(page=page, per_page=per_page)
         elif title:
-            books = Books.query.filter(Books.title.like(title + "%")).all()
+            books = Books.query.filter(Books.title.like(title + "%")).paginate(page=page, per_page=per_page)
         books_list = [book.to_dict() for book in books]
         return jsonify(books_list)
     except Exception as e:
@@ -48,9 +57,9 @@ def post_book():
             Db.session.commit()
             info, status = get_error_details("BOOK_ISSUE_SUCCESS")
             return response_formatter(info, status)
-    # except KeyError as e:
-    #     Db.session.rollback()
-    #     return response_formatter(f"Missing field: {str(e)}", 400)
+    except KeyError as e:
+        Db.session.rollback()
+        return response_formatter(f"Missing field: {str(e)}", 400)
     except Exception as e:
         Db.session.rollback()
         return response_formatter(str(e), 500)
@@ -58,7 +67,8 @@ def post_book():
 
 def get_book(book_id):
     try:
-        req_book = Books.query.get(book_id)
+        page, per_page = get_params_details()
+        req_book = Books.query.get(book_id).paginate(page=page, per_page=per_page)
         if req_book is None:
             info, status = get_error_details("ERR_BOOK_NOT_FOUND")
             return response_formatter(info, status)
@@ -76,6 +86,7 @@ def update_book(book_id):
 
         for key, value in request.json.items():
             if hasattr(req_book, key) and key in book_allowed_fields:
+                print(key, value)
                 setattr(req_book, key, value)
 
         Db.session.commit()
@@ -105,8 +116,9 @@ def delete_book(book_id):
 
 
 def get_members_details():
+    page, per_page = get_params_details()
     try:
-        member_list = Member.query.all()
+        member_list = Member.query.paginate(page=page, per_page=per_page)
         return jsonify([member.to_dict() for member in member_list])
     except Exception as e:
         return response_formatter(str(e), 500)
@@ -124,6 +136,11 @@ def add_member_details():
         user_name = request.json['name']
         email = request.json['email']
         role = request.json['role']
+
+        # validations
+        if role not in ['admin', 'member']:
+            return response_formatter("Invalid role. Must be 'admin' or 'member'", 400)
+
         member_name_list = [member.to_dict()['name'] for member in member_list]
         if is_record_available(member_name_list, user_name):
             info, status = get_error_details("ERR_MEMBER_EXISTS")
@@ -143,10 +160,16 @@ def add_member_details():
 
 
 def get_member_details(member_id):
+    access = False
     try:
         member_details = Member.query.get(member_id)
         if member_details is None:
             info, status = get_error_details("ERR_MEMBER_NOT_FOUND")
+            return response_formatter(info, status)
+
+        # Check if the user's email matches or if they are an admin
+        if member_details.to_dict()['email'] != get_email_from_jwt() and not is_admin():
+            info, status = get_error_details("ERR_UNAUTHORIZED")
             return response_formatter(info, status)
         return jsonify(member_details.to_dict())
     except Exception as e:
@@ -159,17 +182,19 @@ def update_member(member_id):
         if member_details is None:
             info, status = get_error_details("ERR_MEMBER_NOT_FOUND")
             return response_formatter(info, status)
-        else:
-            for key, value in request.json.items():
-                if getattr(member_details, key) and key in member_allowed_properties:
-                    if key == "password":
-                        hashed_password = generate_hashed_password(value)
-                        setattr(member_details, key, hashed_password)
-                    else:
-                        setattr(member_details, key, value)
-            Db.session.commit()
-            info, status = get_error_details("MEMBER_DETAILS_UPDATE_SUCCESS")
+        if member_details.to_dict()['email'] != get_email_from_jwt() and not is_admin():
+            info, status = get_error_details("ERR_UNAUTHORIZED")
             return response_formatter(info, status)
+        for key, value in request.json.items():
+            if getattr(member_details, key) and key in member_allowed_properties:
+                if key == "password":
+                    hashed_password = generate_hashed_password(value)
+                    setattr(member_details, key, hashed_password)
+                else:
+                    setattr(member_details, key, value)
+        Db.session.commit()
+        info, status = get_error_details("MEMBER_DETAILS_UPDATE_SUCCESS")
+        return response_formatter(info, status)
     except AttributeError as e:
         Db.session.rollback()
         return response_formatter(f"Attribute error: {str(e)}", 400)
@@ -181,6 +206,10 @@ def delete_member(member_id):
         if member_details is None:
             info, status = get_error_details("ERR_MEMBER_NOT_FOUND")
             return response_formatter(info, status)
+        if member_details.to_dict()['email'] != get_email_from_jwt() and not is_admin():
+            info, status = get_error_details("ERR_UNAUTHORIZED")
+            return response_formatter(info, status)
+
         Db.session.delete(member_details)
         Db.session.commit()
         info, status = get_error_details("MEMBER_DELETED_SUCCESS")
@@ -219,7 +248,6 @@ def issue_book():
             setattr(book_details, 'book_count', count - 1)
             book_issue = BookMember(book_id=book_id, issue_date=issue_date, member_id=member_id, debt=0,
                                     return_date=None, is_book_returned=False)
-            # raise Exception("Just for fun")
             Db.session.add(book_issue)
             Db.session.commit()
             info, status = get_error_details("BOOK_ISSUE_SUCCESS")
@@ -263,7 +291,8 @@ def get_highest_paying_members():
 def get_famous_books_list():
     #  select book_member.book_id, books.title, count(book_member.member_id) from book_member
     #  left join books on book_member.book_id = books.id group by book_member.book_id;
-    results = Db.session.query(Books.id, Books.title, func.count(BookMember.member_id)).join(BookMember, Books.id== BookMember.book_id).group_by(BookMember.book_id).order_by(func.count(BookMember.member_id).desc()).all()
+    page, per_page = get_params_details()
+    results = Db.session.query(Books.id, Books.title, func.count(BookMember.member_id)).join(BookMember, Books.id== BookMember.book_id).group_by(BookMember.book_id).order_by(func.count(BookMember.member_id).desc()).paginate(page=page, per_page=per_page)
     keys_list = ["id", "title", "count_of_members_taken"]
     serialised_json = map_keys_to_values(keys_list, results)
     return serialised_json
